* ♻️ [refactor] TopViewController -> TripLogSummaryView 분리
